+++ Generator/Source/main.swift
@@ -17,5 +17,10 @@ let helpCommand = HelpCommand(registry: registry)
 registry.register(helpCommand)
 
 registry.main(defaultVerb: helpCommand.verb) { error in
-    fputs(error.description + "\n", stderr)
+    switch error {
+    case .stderrUsed:
+        break
+    default:
+        fputs(error.description + "\n", stderr)
+    }
 }